Update all 'git clone' commands to checkout release branch (#447)
I am merging this PR due to an apparent issue with the CODEOWNER requirement. Xintian Wu is requesting assistance.

"This PR is critical for the timely release. Please check why I cannot force merge this PR? My permission of force merging PRs sometimes works and sometimes does not (don’t see any option this time on the UI.) There is a long list of approvers in this PR and essentially blocked it. If possible, could you please merge it?"

https://opensource.intel.com/jira/servicedesk/customer/portal/1/DIO-20737